+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasetop.py
@@ -100,7 +100,7 @@ class CreateDatasetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-08-31T11:24:22.202Z")
+    updated: Optional[datetime] = parse_datetime("2026-09-01T06:04:38.639Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")