+++ packages/orq-rc/src/orq_ai_sdk/models/get_v2_human_eval_sets_id_op.py
@@ -83,7 +83,7 @@ class GetV2HumanEvalSetsIDResponseBody2(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-08-31T11:24:22.202Z")
+    updated: Optional[datetime] = parse_datetime("2026-09-01T06:04:38.639Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")
@@ -168,7 +168,7 @@ class GetV2HumanEvalSetsIDResponseBody1(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-08-31T11:24:22.202Z")
+    updated: Optional[datetime] = parse_datetime("2026-09-01T06:04:38.639Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")